Separate uploaded file validation logic from the model "Attachments"
